feat: move create cover to cover class
